\name{demostability}

\alias{demostability}

\title{Demo file for three batch}

\description{
  This is a demo file.  First, the stability data is fitted with ANCOVA.
  The result in this example shows that the test rejects the hypothesis of equality of
  slopes (there is a significant difference in slopes among batches).

}

\keyword{misc}